requirements.txt
pymongo[srv]
plotly
pandas
streamlit
python_dotenv